fix(copilot): share workflow cancellation boundary
